Move a couple of dependencies to devDependencies, and run tests on lib/ rather than src/
