+++ lib/index.js
@@ -35,32 +35,8 @@ var _functionArgNames2 = _interopRequireDefault(_functionArgNames);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-var getReducerArgNames = function getReducerArgNames(reducerFunc, actionType) {
-    if (reducerFunc != null) {
-        var reducerArgs = _functionArgNames2.default.getArgs(reducerFunc);
+// --------------------------------------------------------------------  General helpers
 
-        if (reducerArgs.length > 1) {
-            var secondArg = reducerArgs[1];
-            if (secondArg.substr(0, 4) === "_ref") {
-                var refArgs = _functionArgNames2.default.getRefs(reducerFunc, secondArg);
-                if (refArgs == null) {
-                    console.warn("Possible flaw in duck action '" + actionType + "': the reducer function expects a deconstructed object ( e.g. {name1, name2, name3} ) as its " + "second argument, but this seems empty");
-                }
-                return refArgs || [];
-            }
-        }
-    }
-};
-
-var makeReducer = exports.makeReducer = function makeReducer(reducerTable, initialState) {
-    return function () {
-        var state = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : initialState;
-        var action = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : {};
-
-        var reducer = reducerTable[action.type];
-        return reducer ? reducer(state, action) : state;
-    };
-};
 
 var actionNum = 0;
 var getActionType = function getActionType(prefix, actionName) {
@@ -92,7 +68,7 @@ var buildMaps = function buildMaps(prefix, actionAndReducerMap, checkAndWarn, lo
 
         var actionType = getActionType(prefix, actionName);
         var reducerFunction = actionAndReducerMap[actionName];
-        var actionArgumentNames = getReducerArgNames(reducerFunction, actionName) || [];
+        var actionArgumentNames = getReducerArgNames(reducerFunction, actionType) || [];
 
         if (checkAndWarn) {
             (0, _check2.default)(actionType, actionArgumentNames, reducerFunction);
@@ -110,6 +86,37 @@ var buildMaps = function buildMaps(prefix, actionAndReducerMap, checkAndWarn, lo
 
     return [actionCreatorMap, reducerMap, typeMap];
 };
+
+// -------------------------------------------------------------  Reducer-specific helpers
+
+var getReducerArgNames = function getReducerArgNames(reducerFunc, actionType) {
+    if (reducerFunc != null) {
+        var reducerArgs = _functionArgNames2.default.getArgs(reducerFunc);
+
+        if (reducerArgs.length > 1) {
+            var secondArg = reducerArgs[1];
+            if (secondArg.substr(0, 4) === "_ref") {
+                var refArgs = _functionArgNames2.default.getRefs(reducerFunc, secondArg);
+                if (refArgs == null) {
+                    console.warn("Possible flaw in duck action '" + actionType + "': the reducer function expects a deconstructed object ( e.g. {name1, name2, name3} ) as its " + "second argument, but this seems empty");
+                }
+                return refArgs || [];
+            }
+        }
+    }
+};
+
+var makeReducer = exports.makeReducer = function makeReducer(reducerTable, initialState) {
+    return function () {
+        var state = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : initialState;
+        var action = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : {};
+
+        var reducer = reducerTable[action.type];
+        return reducer ? reducer(state, action) : state;
+    };
+};
+
+//------------------------------------------------------------  Entry: class
 
 /**
  *  Creates an action-actioncreator-reducer unified complex: a redux duck.
