+++ package.json
@@ -5,10 +5,10 @@
   "main": "lib/index.js",
   "scripts": {
     "compile": "rimraf lib/* && babel src -d lib",
-    "prepublish": "npm run checkall && npm run compile",
+    "prepublish": "npm run checkall",
     "test": "NODE_PATH=./ mocha --compilers js:babel-core/register --require babel-polyfill --require test/test_helper.js \"test/**/*@(.js)\"",
     "eslint": "eslint ./*.js src/ test/",
-    "checkall": "npm run eslint & npm run test"
+    "checkall": "npm run eslint && npm run compile && npm run test"
   },
   "repository": {
     "type": "git",
@@ -21,9 +21,7 @@
   },
   "homepage": "https://github.com/espen42/duckfactory#readme",
   "dependencies": {
-    "babel-preset-stage-0": "^6.22.0",
-    "babel-runtime": "^6.23.0",
-    "deep-freeze": "^0.0.1"
+    "babel-runtime": "^6.23.0"
   },
   "devDependencies": {
     "babel-cli": "^6.23.0",
@@ -32,7 +30,9 @@
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-polyfill": "6.20.0",
     "babel-preset-es2015": "^6.22.0",
+    "babel-preset-stage-0": "^6.22.0",
     "chai": "3.5.0",
+    "deep-freeze": "^0.0.1",
     "eslint": "3.14.0",
     "eslint-plugin-react": "^6.10.0",
     "jsdom": "^9.11.0",
