+++ test/duckfactory_spec.js
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import deepFreeze from 'deep-freeze';
 
-import DuckFactory from '../src';
+import DuckFactory from '../lib';
